packages/cli: move reading of plugin template data to plugin diff
