+++ packages/cli/src/commands/plugin/diff.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+import fs from 'fs-extra';
 import { Command } from 'commander';
 import {
   diffTemplateFiles,
@@ -23,6 +24,13 @@ import {
   makeCheckPromptFunc,
   yesPromptFunc,
 } from '../../lib/diff';
+import { paths } from '../../lib/paths';
+import { version } from '../../lib/version';
+
+export type PluginData = {
+  id: string;
+  name: string;
+};
 
 const fileHandlers = [
   {
@@ -54,7 +62,36 @@ export default async (cmd: Command) => {
     promptFunc = yesPromptFunc;
   }
 
-  const templateFiles = await diffTemplateFiles('default-plugin');
+  const data = await readPluginData();
+  const templateFiles = await diffTemplateFiles('default-plugin', {
+    version,
+    ...data,
+  });
   await handleAllFiles(fileHandlers, templateFiles, promptFunc);
   await finalize();
 };
+
+// Reads templating data from the existing plugin
+async function readPluginData(): Promise<PluginData> {
+  let name: string;
+  try {
+    const pkg = require(paths.resolveTarget('package.json'));
+    name = pkg.name;
+  } catch (error) {
+    throw new Error(`Failed to read target package, ${error}`);
+  }
+
+  const pluginTsContents = await fs.readFile(
+    paths.resolveTarget('src/plugin.ts'),
+    'utf8',
+  );
+  // TODO: replace with some proper parsing logic or plugin metadata file
+  const pluginIdMatch = pluginTsContents.match(/id: ['"`](.+?)['"`]/);
+  if (!pluginIdMatch) {
+    throw new Error(`Failed to parse plugin.ts, no plugin ID found`);
+  }
+
+  const id = pluginIdMatch[1];
+
+  return { id, name };
+}
